Replace getPageContext with theme
