--- lib/getPageContext.js
@@ -1,55 +1,0 @@
-/* eslint-disable no-underscore-dangle */
-
-import { indigo, teal } from '@material-ui/core/colors';
-import {
-  createGenerateClassName,
-  createMuiTheme,
-} from '@material-ui/core/styles';
-import { SheetsRegistry } from 'jss';
-
-// A theme with custom primary and secondary color.
-// It's optional.
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: teal[400],
-      main: teal[600],
-      dark: teal[800],
-    },
-    secondary: {
-      light: indigo[400],
-      main: indigo[600],
-      dark: indigo[800],
-    },
-  },
-  typography: {
-    useNextVariants: true,
-  },
-});
-
-function createPageContext() {
-  return {
-    theme,
-    // This is needed in order to deduplicate the injection of CSS in the page.
-    sheetsManager: new Map(),
-    // This is needed in order to inject the critical CSS.
-    sheetsRegistry: new SheetsRegistry(),
-    // The standard class name generator.
-    generateClassName: createGenerateClassName(),
-  };
-}
-
-export default function getPageContext() {
-  // Make sure to create a new context for every server-side request so that
-  // data isn't shared between connections (which would be bad).
-  if (!process.browser) {
-    return createPageContext();
-  }
-
-  // Reuse context on the client-side.
-  if (!global.__INIT_MATERIAL_UI__) {
-    global.__INIT_MATERIAL_UI__ = createPageContext();
-  }
-
-  return global.__INIT_MATERIAL_UI__;
-}
